Remove legacy `--runtime=nvidia` to avoid failures
This causes a failure unless the legacy package nvidia-docker2 is also installed. The correct option `--gpus=all` was already present, which resulted in correct behavior **only** if a recent version of docker was installed **alongside** the legacy package `nvidia-docker2` (e.g. when Docker was updated from a pre-19.03 version, but `nvidia-docker2` was not subsequently removed).

The correct option to use `--gpus=all` was already present